Final DON
Bench: 2063964

+++ src/misc.h
@@ -142,8 +142,8 @@ __extension__ using u128 = unsigned __int128;
 __extension__ using i128 = signed __int128;
 #endif
 
-using NumaIndex = usize;
-using CpuIndex  = usize;
+using NumaIndex = u16;
+using CpuIndex  = u16;
 
 using Strings     = std::vector<std::string>;
 using StringViews = std::vector<std::string_view>;
